refactor(profile): Move relevant code to its feature folder
